Playing around to see what editor state is needed.
Had a slight detour trying to work out the best
way to deal with c strings that we need to pass
into Raylib functions.  Strings constants are
null terminated by default, so we've not noticed
problems passing those in.  But some raylib functions
like GuiTextBox can modify the length of a null
terminated string, which requires extra care.

I'm probably mis-using the null-terminated notation
for slices.