Fix bug with tile coordinates being off by 1/2 tile.
Oddly, with DrawTexturePro, your destRect (x,y) should be where
you want the center of the image drawn.  Which is really weird,
since in other contexts with a Rectangle, it's the top left of
the rectangle.  Really bad API, I had to stare at the example
code for a while to get it.